+++ src/main/scala/tscfg/scalaGenerator.scala
@@ -62,7 +62,10 @@ object scalaGenerator {
         // </recurse>
 
         // <apply>
-        out.println(s"$indent  def apply(c: scala.Option[$TypesafeConfigClassName]): $className = {")
+        // #13: add `build` auxiliary method to avoid type erasure issue in case of optional member
+        out.println(s"$indent  def apply(c: com.typesafe.config.Config) = build(Some(c))")
+        out.println(s"$indent  def apply() = build(None)")
+        out.println(s"$indent  def build(c: scala.Option[$TypesafeConfigClassName]): $className = {")
         out.println(s"$indent    $className(")
 
         comma = indent
@@ -74,7 +77,7 @@ object scalaGenerator {
 
             case BranchNode(k, _)  =>
               val className = upperFirst(k.simple)
-              out.print(s"""      $className(c.map(c => if (c.hasPath("${k.simple}")) Some(c.getConfig("${k.simple}")) else None).get)""")
+              out.print(s"""      $className.build(c.map(c => if (c.hasPath("${k.simple}")) Some(c.getConfig("${k.simple}")) else None).get)""")
           }
           comma = s",\n$indent"
         }